Why does git commit require a message?
