+++ snakecord/states/memberstate.py
@@ -89,40 +89,3 @@ class GuildMemberState(BaseState):
         await rest.remove_guild_member.request(
             session=self.manager.rest,
             fmt=dict(guild_id=self.guild.id, user_id=user_id))
-
-    async def fetch_prune_count(self, days=None, include_roles=None):
-        params = {}
-
-        if days is not None:
-            params['days'] = int(days)
-
-        if include_roles is not None:
-            include_roles = {
-                str(Snowflake.try_snowflake(r)) for r in include_roles
-            }
-            params['include_roles'] = ','.join(include_roles)
-
-        data = await rest.get_guild_prune_count.request(
-            session=self._state.manager.rest,
-            fmt=dict(guild_id=self.guild.id))
-
-        return data['pruned']
-
-    async def prune(self, **kwargs):
-        keys = rest.begin_guild_prune.json
-
-        try:
-            roles = Snowflake.try_snowflake_set(kwargs['roles'])
-            kwargs['roles'] = tuple(roles)
-        except KeyError:
-            pass
-
-        _validate_keys(f'{self.__class__.__name__}.begin_prune',
-                       kwargs, (), keys)
-
-        data = await rest.begin_guild_prune.request(
-            session=self._state.manager.rest,
-            fmt=dict(guild_id=self.guild.id),
-            json=kwargs)
-
-        return data['pruned']
